remove old instance slice, start porting components
