+++ src/shell/components/global-tabs/GlobalTabs.js
@@ -4,13 +4,11 @@ import { useHistory } from "react-router-dom";
 import cx from "classnames";
 import usePrevious from "react-use/lib/usePrevious";
 import debounce from "lodash/debounce";
-
 import { AppLink } from "@zesty-io/core/AppLink";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimesCircle } from "@fortawesome/free-solid-svg-icons";
-
 import { closeTab, openTab, loadTabs, rebuildTabs } from "shell/store/ui";
-
+import { useGetInstanceQuery } from "shell/services/accounts";
 import styles from "./GlobalTabs.less";
 
 const MIN_TAB_WIDTH = 150;
@@ -19,10 +17,10 @@ const TAB_PADDING = 16;
 const TAB_BORDER = 1;
 
 export default React.memo(function GlobalTabs() {
+  const { data: instance } = useGetInstanceQuery();
   const history = useHistory();
   const dispatch = useDispatch();
   const tabs = useSelector(state => state.ui.tabs);
-  const instanceZUID = useSelector(state => state.instance.ZUID);
   const loadedTabs = useSelector(state => state.ui.loadedTabs);
   const prevPath = usePrevious(history.location.pathname);
   const [windowWidth, setWindowWidth] = useState(window.innerWidth);
@@ -44,8 +42,8 @@ export default React.memo(function GlobalTabs() {
 
   // load tabs from Indexeddb
   useEffect(() => {
-    dispatch(loadTabs(instanceZUID));
-  }, [instanceZUID]);
+    dispatch(loadTabs(instance.ZUID));
+  }, [instance.ZUID]);
 
   // openTab every time path changes
   useEffect(() => {
